Introducing unique ids for node slot configurations. Hub injects UID of the matched configuration to the new request payload. This allows to avoid secondary matching on the node, it knows immediately what slot configuration to use.
